moves to lib folder
